Add a mostly stubbed out CollectionListView
And update base CollectionDetailView and CollectionListView
to enable the base generic put/patch/post.

In theory, if collection import is running elsewhere, the
POST /collections/ doesn't need to get pass a artifact tarfile
and process it. If the import happens elsewhere, then maybe
can just POST the data to create the collection. maybe.