[wip] Add UiCollection* for of /_ui/collections schema
This is first pass at 'as compatible to community
galaxy /api/internal as possible'. It has not yet
been whittled  down to stuff that just makes sense
for hub.

Note not quite right at the moment, get /_ui/collections returns
a UiCollection with 'latest_version' as a list of
VersionDetailSummary and get /ui/collections/{namespace}/{name}
returns a UiCollection-ish object with 'latest_version' as a
list of VersionDetail

Issue: ansible/galaxy-dev#55
Signed-off-by: Adrian Likins <alikins@redhat.com>